+++ .storybook/decorators/PreviewHTML/PreviewHTML.tsx
@@ -1,6 +1,8 @@
 import React, { useEffect, useState } from "react";
+import { DARK_MODE_EVENT_NAME, useDarkMode } from "storybook-dark-mode";
+import { addons } from "@storybook/preview-api";
+import { StoryContext, StoryFn } from "@storybook/react";
 
-import { useDarkMode } from "storybook-dark-mode";
 import {
   createElementWithClasses,
   processCode,
@@ -9,13 +11,18 @@ import {
 } from "./previewHTML.helpers";
 
 import "./style.css";
-import { StoryContext, StoryFn } from "@storybook/react";
+
+const channel = addons.getChannel();
 
 export const PreviewHTML = (Story: StoryFn, context: StoryContext) => {
-  const [darkMode, setDarkMode] = useState(false);
+  const [darkMode, setDarkMode] = useState(useDarkMode());
+
+  useEffect(() => {
+    channel.on(DARK_MODE_EVENT_NAME, setDarkMode);
+    return () => channel.removeListener(DARK_MODE_EVENT_NAME, setDarkMode);
+  }, [channel, setDarkMode]);
 
   const changeDarkMode = (message: MessageEvent) => {
-    console.log(`Received ${message}`);
     if (message.data?.type === "change_dark_theme") {
       setDarkMode(message.data.value);
     }
@@ -42,26 +49,7 @@ export const PreviewHTML = (Story: StoryFn, context: StoryContext) => {
     mainWrapper.appendChild(copyButtonContainer);
 
     const codeContainer = createElementWithClasses("div", "css-8ycahn");
-    const scroller = createElementWithClasses("div", "scroller");
-    const scrollerInner = createElementWithClasses("div", "scroller-inner");
 
-    mainWrapper.addEventListener("mouseenter", () => {
-      scroller.style.position = "absolute";
-      scroller.style.bottom = "2px";
-      scrollerInner.style.width = codeContainer.scrollWidth + "px";
-      scroller.style.width = codeContainer.clientWidth + "px";
-      scrollerInner.style.height = "1px";
-      scroller.appendChild(scrollerInner);
-      scroller.style.overflowX = "scroll";
-      scroller.addEventListener("scroll", () => {
-        codeContainer.scrollLeft = scroller.scrollLeft;
-      });
-
-      mainWrapper.appendChild(scroller);
-    });
-    mainWrapper.addEventListener("mouseleave", () => {
-      mainWrapper.removeChild(scroller);
-    });
     mainWrapper.appendChild(codeContainer);
 
     const htmlDisplay = createElementWithClasses(