+++ packages/implementations/react/src/atomic/Divider/src/divider.stories.tsx
@@ -3,8 +3,6 @@ import { argTypesConvert } from ".storybook/utils";
 import { Divider } from "./Divider";
 import docs from "./divider.docs.json";
 
-console.log("argTypesConvert(docs)", argTypesConvert(docs));
-
 const meta: Meta<typeof Divider> = {
   title: "Atomic/Divider",
   component: Divider,