+++ packages/implementations/react/src/theme/Provider/src/contexts/Manager/Manager.tsx
@@ -1,0 +1,6 @@
+import { createContext } from "react";
+import { ThemeProviderManagerProps } from "./manager.types";
+
+export const ThemeProviderManager = createContext<ThemeProviderManagerProps>(
+  {} as ThemeProviderManagerProps,
+);