+++ packages/implementations/react/src/theme/Provider/src/hooks/useTheme/useTheme.ts
@@ -1,0 +1,9 @@
+import { useContext } from "react";
+
+import {
+  ThemeProviderManager,
+  ThemeProviderManagerProps,
+} from "../../contexts";
+
+export const useTheme: () => ThemeProviderManagerProps = () =>
+  useContext(ThemeProviderManager);