+++ packages/implementations/react/src/theme/Provider/src/index.ts
@@ -1,0 +1,6 @@
+import { ThemeProvider } from "./Provider";
+
+export { ThemeProvider } from "./Provider";
+export { useTheme } from "./hooks";
+export type { ProviderProps as ThemeProviderProps } from "./provider.types";
+export default ThemeProvider;