+++ packages/implementations/react/src/theme/Provider/src/provider.definitions.ts
@@ -1,0 +1,6 @@
+import { variablesDark } from "@rarui/styles";
+
+export const themes = {
+  dark: variablesDark,
+  base: "",
+};