+++ packages/implementations/react/src/theme/Provider/src/provider.types.ts
@@ -1,0 +1,9 @@
+import { HTMLAttributes } from "react";
+
+export type Theme = "dark" | "base";
+
+export interface ProviderProperties {
+  theme?: Theme;
+}
+
+export type ProviderProps = ProviderProperties & HTMLAttributes<HTMLElement>;