+++ packages/implementations/vuejs/.storybook/main.ts
@@ -1,9 +1,10 @@
+import main from "../../../../.storybook/main";
 import type { StorybookConfig } from "@storybook/vue3-vite";
 import path from "path";
 
 export const convertTsConfigPathsToWebpackAliases = () => {
   const rootDir = path.resolve(__dirname, "../");
-  const paths = [];
+  const paths: any = [];
 
   paths["@rarui/tokens"] = path.join(rootDir, "../../../packages/tokens");
   paths["@rarui/styles"] = path.join(
@@ -14,30 +15,29 @@ export const convertTsConfigPathsToWebpackAliases = () => {
     rootDir,
     "../../../packages/core/webpack/src/index.ts",
   );
-
-  console.log("paths", paths);
   return paths;
 };
 
 const config: StorybookConfig = {
-  stories: [
-    "../src/**/*.mdx",
-    "../src/**/*.stories.@(js|jsx|mjs|ts|tsx)",
-    "../src/*/**/*.stories.@(js|jsx|ts|tsx)",
+  staticDirs: [
+    {
+      from: "../../../../.storybook/static",
+      to: "/static",
+    },
   ],
-  addons: [
-    "@storybook/addon-links",
-    "@storybook/addon-essentials",
-    "@storybook/addon-interactions",
-  ],
+  stories: ["../src/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
+  addons: main.addons,
+  docs: main.docs,
   framework: {
     name: "@storybook/vue3-vite",
     options: {},
   },
-  docs: {
-    autodocs: "tag",
-  },
-  async viteFinal(config, options) {
+
+  async viteFinal(config) {
+    config.build = {
+      ...config.build,
+      chunkSizeWarningLimit: 1000,
+    };
     config.resolve = {
       ...config.resolve,
       alias: {
@@ -45,7 +45,11 @@ const config: StorybookConfig = {
         ...convertTsConfigPathsToWebpackAliases(),
       },
     };
+    config.plugins = config.plugins?.filter(
+      (plugin: any) => plugin.name !== "vite:dts",
+    );
     return config;
   },
 };
+
 export default config;