+++ packages/implementations/vuejs/.storybook/preview.ts
@@ -1,13 +1,10 @@
+import basePreview from "../../../../.storybook/preview";
+
 import type { Preview } from "@storybook/vue3";
 
 const preview: Preview = {
   parameters: {
-    controls: {
-      matchers: {
-        color: /(background|color)$/i,
-        date: /Date$/i,
-      },
-    },
+    ...basePreview.parameters,
   },
 };