+++ packages/styles/src/packages/atomic/iconButton/rarui-iconButton.css.ts
@@ -32,7 +32,7 @@ const iconButton = recipe({
     },
     ":focus-within": {
       outline: "none",
-      boxShadow: varsThemeBase.utils["focus-ring"],
+      boxShadow: varsThemeBase.elevation["focus-ring"],
     },
     ":after": {
       transition: "opacity 150ms ease",