+++ packages/styles/src/packages/atomic/input/rarui-input.css.ts
@@ -21,7 +21,7 @@ const container = recipe({
     },
     ":focus-within": {
       outline: "none",
-      boxShadow: varsThemeBase.utils["focus-ring"],
+      boxShadow: varsThemeBase.elevation["focus-ring"],
     },
     selectors: {
       "&:has(input:disabled)": {