+++ packages/styles/src/packages/composite/sideNavigation/rarui-sideNavigation.css.ts
@@ -32,7 +32,7 @@ const sideNavigation = recipe({
     },
     ":focus-within": {
       outline: "none",
-      boxShadow: varsThemeBase.utils["focus-ring"],
+      boxShadow: varsThemeBase.elevation["focus-ring"],
     },
     ":disabled": {
       backgroundColor: varsThemeBase.colors.surface.disabled,